update phase 4 tests
